# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API documentation for E-Wallet Backend",
        "title": "E-Wallet API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/wallet/dashboard": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get authenticated user wallet summary data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wallet"
                ],
                "summary": "Wallet Dashboard",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_bernaddwiki_koda-b7-weekly10_internal_dto.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_bernaddwiki_koda-b7-weekly10_internal_dto.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_bernaddwiki_koda-b7-weekly10_internal_dto.Response"
                        }
                    }
                }
            }
        },
        "/wallet/top-up": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a wallet top-up request using payment method",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wallet"
                ],
                "summary": "Create Top Up",
                "parameters": [
                    {
                        "description": "Top Up Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_bernaddwiki_koda-b7-weekly10_internal_dto.CreateTopUpRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_bernaddwiki_koda-b7-weekly10_internal_dto.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_bernaddwiki_koda-b7-weekly10_internal_dto.Response"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/github_com_bernaddwiki_koda-b7-weekly10_internal_dto.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_bernaddwiki_koda-b7-weekly10_internal_dto.Response"
                        }
                    }
                }
            }
        },
        "/wallet/transaction-report": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get daily transaction totals for authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wallet"
                ],
                "summary": "Transaction Report",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 7,
                        "description": "Number of days to include in report",
                        "name": "days",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "both",
                        "description": "Filter by flow: income, expense, or both",
                        "name": "flow",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_bernaddwiki_koda-b7-weekly10_internal_dto.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_bernaddwiki_koda-b7-weekly10_internal_dto.Response"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/github_com_bernaddwiki_koda-b7-weekly10_internal_dto.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_bernaddwiki_koda-b7-weekly10_internal_dto.Response"
                        }
                    }
                }
            }
        },
        "/wallet/transactions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get paginated transaction history for authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wallet"
                ],
                "summary": "Transaction History",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search term for transaction history",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Limit per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_bernaddwiki_koda-b7-weekly10_internal_dto.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_bernaddwiki_koda-b7-weekly10_internal_dto.Response"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/github_com_bernaddwiki_koda-b7-weekly10_internal_dto.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_bernaddwiki_koda-b7-weekly10_internal_dto.Response"
                        }
                    }
                }
            }
        },
        "/wallet/transfer": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a wallet transfer to another user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wallet"
                ],
                "summary": "Create Transfer",
                "parameters": [
                    {
                        "description": "Transfer Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_bernaddwiki_koda-b7-weekly10_internal_dto.CreateTransferRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_bernaddwiki_koda-b7-weekly10_internal_dto.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_bernaddwiki_koda-b7-weekly10_internal_dto.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_bernaddwiki_koda-b7-weekly10_internal_dto.Response"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/github_com_bernaddwiki_koda-b7-weekly10_internal_dto.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_bernaddwiki_koda-b7-weekly10_internal_dto.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "github_com_bernaddwiki_koda-b7-weekly10_internal_dto.CreateTopUpRequest": {
            "type": "object",
            "required": [
                "amount",
                "payment_method_id"
            ],
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "note": {
                    "type": "string"
                },
                "payment_method_id": {
                    "type": "integer"
                }
            }
        },
        "github_com_bernaddwiki_koda-b7-weekly10_internal_dto.CreateTransferRequest": {
            "type": "object",
            "required": [
                "amount",
                "pin",
                "receiver_id"
            ],
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "note": {
                    "type": "string"
                },
                "pin": {
                    "type": "string"
                },
                "receiver_id": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_bernaddwiki_koda-b7-weekly10_internal_dto.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}